pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "epi-recorder"
version = "2.2.0"
description = "The Flight Recorder for AI Agents. Debug LangChain & CrewAI with execution tracing."
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "EPI Labs", email = "mohdibrahim@epilabs.org"}
]
maintainers = [
    {name = "Mohd Ibrahim Afridi", email = "mohdibrahim@epilabs.org"}
]
keywords = [
    "ai", 
    "debugging", 
    "agents", 
    "langchain", 
    "crewai", 
    "devtools",
    "observability", 
    "llm", 
    "openai", 
    "gemini", 
    "tracing",
    "flight-recorder"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Debuggers",
    "Topic :: Software Development :: Testing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Logging",
    "Typing :: Typed",
    "Framework :: Pydantic",
    "Framework :: Pydantic :: 2"
]

dependencies = [
    "pydantic>=2.0.0",
    "cryptography>=41.0.0",
    "cbor2>=5.6.0",
    "typer[all]>=0.12.0",
    "rich>=13.0.0",
    "google-generativeai>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]

[project.scripts]
epi = "epi_cli.main:app"

[project.urls]
Homepage = "https://epilabs.org"
Documentation = "https://epilabs.org/docs"
Repository = "https://github.com/mohdibrahimaiml/EPI-V2.1.2"
Issues = "https://github.com/mohdibrahimaiml/EPI-V2.1.2/issues"
Discussions = "https://github.com/mohdibrahimaiml/EPI-V2.1.2/discussions"

[tool.setuptools.packages.find]
where = ["."]
include = ["epi_core*", "epi_cli*", "epi_recorder*", "epi_analyzer*", "epi_viewer_static*", "epi_postinstall*"]

[tool.setuptools]
py-modules = ["epi_postinstall"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=epi_core --cov=epi_cli --cov-report=term-missing --cov-report=html"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

